rename: recoil 상태들 atom과 selector로 분류
