--- src/recoil/selectedCardItems.ts
@@ -1,58 +1,0 @@
-import { atomFamily, selector, selectorFamily } from 'recoil';
-import { recoilPersist } from 'recoil-persist';
-import { cartItemsState } from './cartItems';
-import { CartItemType } from '../type';
-
-const { persistAtom } = recoilPersist({
-  key: 'isCartItemSelected',
-  storage: localStorage,
-});
-
-export const isCartItemSelectedState = atomFamily<boolean, number>({
-  key: 'isCartItemSelected',
-  default: true,
-  effects_UNSTABLE: [persistAtom],
-});
-
-export const selectedCartItemsState = selector<CartItemType[]>({
-  key: 'selectedCartItemsState',
-  get: ({ get }) => {
-    return get(cartItemsState).filter((cartItem) =>
-      get(isCartItemSelectedState(cartItem.id)),
-    );
-  },
-});
-
-export const isAllCartItemSelectedState = selectorFamily<boolean, number[]>({
-  key: 'selectedAllCartItem',
-  get:
-    (cartItemIds) =>
-    ({ get }) => {
-      return cartItemIds.every((itemId) =>
-        get(isCartItemSelectedState(itemId)),
-      );
-    },
-  set:
-    (cartItemIds) =>
-    ({ set }, newValue) => {
-      cartItemIds.forEach((itemId) => {
-        set(isCartItemSelectedState(itemId), newValue);
-      });
-    },
-});
-
-export const isSomeCartItemSelectedState = selector<boolean>({
-  key: 'selectedSomeCartItem',
-  get: ({ get }) => {
-    return get(cartItemIdsState).some((cartItemId) =>
-      get(isCartItemSelectedState(cartItemId)),
-    );
-  },
-});
-
-export const cartItemIdsState = selector({
-  key: 'CartItemIds',
-  get: ({ get }) => {
-    return get(cartItemsState).map((cartItem) => cartItem.id);
-  },
-});
